First commit Flask-Mircoblog-Reference implementation
This web-application is written in Python and the Flask extension using
the book Flask-Mega-Tutorial by Migual Grinberg.

Some remarks:
- the language transformation of boilerplate is skipped.
- blog post transformation is enabled
- configurations is done by json-files instead of environment vars

Future changes:

The environment JSON files are injected by giving the paths as
arguments; you can do this with Docker files in production.
However the FLASK-tool ($ flask run) does not support user
arguments. So a future change will use env-file to support the paths to
the JSON files.